Do type casts

+++ EnvironmentSimulator/code-examples/follow_reference/follow_reference.cpp
@@ -138,12 +138,16 @@ int main(int argc, char* argv[])
         {
             // find lateral distance between Ego and reference vehicle x-axis
             double front_axle[2] = {0.0, 0.0};
-            if (ego_state.speed >= 0.0)
+            if (ego_state.speed >= 0.0f)
             {
                 // measure from Ego front axle when driving forward
                 rotate_vec(x_offset, 0.0, ego_state.h, front_axle[0], front_axle[1]);
             }
-            double dy = distance_point_to_line(ego_state.x + front_axle[0], ego_state.y + front_axle[1], ref_state.x, ref_state.y, ref_state.h);
+            double dy = distance_point_to_line(static_cast<double>(ego_state.x) + front_axle[0],
+                                               static_cast<double>(ego_state.y) + front_axle[1],
+                                               ref_state.x,
+                                               ref_state.y,
+                                               ref_state.h);
 
             // find longitudinal distance and heading difference between Ego and reference vehicle
             double dx = distance_point_to_line(ego_state.x, ego_state.y, ref_state.x, ref_state.y, static_cast<double>(ref_state.h) + M_PI_2);
